apps/search/requirements.txt
django==5.0.6
flake8==7.0.0
python-dotenv==1.0.1
django-split-settings==1.3.1
django-cors-headers==4.3.1
uwsgi==2.0.25.1
requests==2.32.3
isort==5.13.2
pydantic-settings==2.3.3
psycopg2-binary==2.9.9